refactor: expose `JwtKey` through `test-utils` feature
We want to reuse key generation helpers in the e2e integration test. For
that we have to expose it. We gate it behind test-utils feature to not
expose it to the general API.
